starting test fixes, code cleanup
