+++ frontend/src/utils/validation/zodValidation.ts
@@ -1,6 +1,13 @@
-import dayjs from "dayjs";
 import { FrontendErrorDetails } from "src/types/apiResponseTypes";
+
 import { z } from "zod";
+
+import { isFieldInSchema, unwrapSchema } from "./zodFormData";
+
+type ValidationTranslator = {
+  has: (key: string) => boolean;
+  (key: string): string;
+};
 
 export function getValidationTypeFromZodIssue(
   issue: z.ZodIssue,
@@ -39,113 +46,6 @@ export function getValidationTypeFromZodIssue(
   }
 }
 
-type FormDataAdapters = Record<string, (formData: FormData) => unknown>;
-
-export function formDataToZodInput<T extends z.ZodRawShape>(
-  formData: FormData,
-  schema: z.ZodObject<T>,
-  adapters: FormDataAdapters = {},
-) {
-  const result: Record<string, unknown> = {};
-
-  for (const [fieldName, fieldSchema] of Object.entries(schema.shape)) {
-    const adapter = adapters[fieldName];
-
-    if (adapter) {
-      result[fieldName] = adapter(formData);
-      continue;
-    }
-
-    const rawValue = formData.get(fieldName);
-    result[fieldName] = normalizeValueForSchema(rawValue, fieldSchema);
-  }
-
-  return result;
-}
-
-function normalizeValueForSchema(
-  value: FormDataEntryValue | null,
-  schema: z.ZodTypeAny,
-): unknown {
-  const { schema: unwrappedSchema, nullable, optional } = unwrapSchema(schema);
-
-  if (value === null) {
-    if (nullable) return null;
-    if (optional) return undefined;
-    return undefined;
-  }
-
-  if (typeof value !== "string") {
-    return value;
-  }
-
-  if (value === "") {
-    if (nullable) {
-      return null;
-    }
-
-    // Important: preserve "" for required strings/dates so Zod
-    // can report an actual validation failure.
-    return "";
-  }
-
-  if (unwrappedSchema instanceof z.ZodNumber) {
-    const normalized = value.replace(/[$,\s]/g, "");
-    return Number(normalized);
-  }
-
-  if (unwrappedSchema instanceof z.ZodBoolean) {
-    if (value === "true") return true;
-    if (value === "false") return false;
-  }
-
-  return value;
-}
-
-function unwrapSchema(schema: z.ZodTypeAny): {
-  schema: z.ZodTypeAny;
-  nullable: boolean;
-  optional: boolean;
-} {
-  let current = schema;
-  let nullable = false;
-  let optional = false;
-
-  while (true) {
-    if (current instanceof z.ZodNullable) {
-      nullable = true;
-      current = current.unwrap();
-      continue;
-    }
-
-    if (current instanceof z.ZodOptional) {
-      optional = true;
-      current = current.unwrap();
-      continue;
-    }
-
-    break;
-  }
-
-  return {
-    schema: current,
-    nullable,
-    optional,
-  };
-}
-
-export function isFieldInSchema<T extends z.ZodRawShape>(
-  schema: z.ZodObject<T>,
-  field: string,
-): field is Extract<keyof T, string> {
-  return field in schema.shape;
-}
-
-type ValidationTranslator = {
-  has: (key: string) => boolean;
-  (key: string): string;
-};
-
 export function getTranslatedValidationMessage(
   fieldTranslations: ValidationTranslator,
   genericTranslations: ValidationTranslator,
@@ -168,6 +68,54 @@ export function getTranslatedValidationMessage(
   }
 
   return fallbackMessage;
+}
+
+export function getZodValidationErrors<TShape extends z.ZodRawShape>(
+  error: z.ZodError,
+  validationData: Record<string, unknown>,
+  schema: z.ZodObject<TShape>,
+  fieldTranslations: ValidationTranslator,
+  genericTranslations: ValidationTranslator,
+  field?: string,
+): Partial<Record<Extract<keyof TShape, string>, string[]>> {
+  const validationErrors: Partial<
+    Record<Extract<keyof TShape, string>, string[]>
+  > = {};
+
+  for (const issue of error.issues) {
+    const issueField = issue.path[0]?.toString();
+
+    if (!issueField || !isFieldInSchema(schema, issueField)) {
+      continue;
+    }
+
+    if (field && issueField !== field) {
+      continue;
+    }
+
+    const fieldSchema = schema.shape[issueField];
+
+    const validationType = getValidationTypeFromZodIssue(
+      issue,
+      validationData[issueField],
+      fieldSchema,
+    );
+
+    const message = getTranslatedValidationMessage(
+      fieldTranslations,
+      genericTranslations,
+      issueField,
+      validationType,
+      issue.message,
+    );
+
+    validationErrors[issueField] = [
+      ...(validationErrors[issueField] ?? []),
+      message,
+    ];
+  }
+
+  return validationErrors;
 }
 
 export function mapApiValidationErrors<TShape extends z.ZodRawShape>(
@@ -199,7 +147,10 @@ export function mapApiValidationErrors<TShape extends z.ZodRawShape>(
     );
 
     if (field && isFieldInSchema(schema, field)) {
-      validationErrors[field] = [...(validationErrors[field] ?? []), message];
+      validationErrors[field] = [
+        ...(validationErrors[field] ?? []),
+        message,
+      ];
     } else {
       unmappedMessages.push(message);
     }
@@ -216,111 +167,4 @@ export function mapApiValidationErrors<TShape extends z.ZodRawShape>(
           ? undefined
           : response.message || genericMessage,
   };
-}
-
-export function normalizeDateValue(value: string | null): string | null {
-  if (!value) {
-    return value;
-  }
-
-  const parsed = dayjs(value, ["MM/DD/YYYY", "YYYY-MM-DD"], true);
-
-  return parsed.isValid() ? parsed.format("YYYY-MM-DD") : value;
-}
-
-export function getZodValidationMessages<TShape extends z.ZodRawShape>(
-  error: z.ZodError,
-  validationData: Record<string, unknown>,
-  schema: z.ZodObject<TShape>,
-  fieldTranslations: ValidationTranslator,
-  genericTranslations: ValidationTranslator,
-  field?: string,
-): Partial<Record<Extract<keyof TShape, string>, string[]>> {
-  const validationErrors: Partial<
-    Record<Extract<keyof TShape, string>, string[]>
-  > = {};
-
-  for (const issue of error.issues) {
-    const issueField = issue.path[0]?.toString();
-
-    if (!issueField || !isFieldInSchema(schema, issueField)) {
-      continue;
-    }
-
-    if (field && issueField !== field) {
-      continue;
-    }
-
-    const fieldSchema = schema.shape[issueField];
-
-    const validationType = getValidationTypeFromZodIssue(
-      issue,
-      validationData[issueField],
-      fieldSchema,
-    );
-
-    const message = getTranslatedValidationMessage(
-      fieldTranslations,
-      genericTranslations,
-      issueField,
-      validationType,
-      issue.message,
-    );
-
-    validationErrors[issueField] = [
-      ...(validationErrors[issueField] ?? []),
-      message,
-    ];
-  }
-
-  return validationErrors;
-}
-
-export function getZodValidationErrors<TShape extends z.ZodRawShape>(
-  error: z.ZodError,
-  validationData: Record<string, unknown>,
-  schema: z.ZodObject<TShape>,
-  fieldTranslations: ValidationTranslator,
-  genericTranslations: ValidationTranslator,
-  field?: string,
-): Partial<Record<Extract<keyof TShape, string>, string[]>> {
-  const validationErrors: Partial<
-    Record<Extract<keyof TShape, string>, string[]>
-  > = {};
-
-  for (const issue of error.issues) {
-    const issueField = issue.path[0]?.toString();
-
-    if (!issueField || !isFieldInSchema(schema, issueField)) {
-      continue;
-    }
-
-    // If a field was supplied, only collect errors for that field.
-    if (field && issueField !== field) {
-      continue;
-    }
-
-    const fieldSchema = schema.shape[issueField];
-
-    const validationType = getValidationTypeFromZodIssue(
-      issue,
-      validationData[issueField],
-      fieldSchema,
-    );
-
-    const message = getTranslatedValidationMessage(
-      fieldTranslations,
-      genericTranslations,
-      issueField,
-      validationType,
-      issue.message,
-    );
-
-    validationErrors[issueField] = [
-      ...(validationErrors[issueField] ?? []),
-      message,
-    ];
-  }
-
-  return validationErrors;
 }
